parse: Fix behavior when --fix-dates is not specified
Previously, if --fix-dates was not specified, it would behave as if
--fix-dates='monthfirst' were used.

This change was inadvertently left out of e82d3148f569572fe45ca02913604d43e5b8ff5e.
Previous code ran without errors because `fix_dates` is the name of a
function.

The bug was noticed by a Mypy warning hinting that `fix_dates` will
always be truthy.