titer_model: Remove top-level entrypoint
Once upon a time, this Python file might have been used as a standalone
program. Nowadays, it is only used by titers.py. This means the code
removed in this commit should be unused.

Removal of the entrypoint code allows the removal of a Mypy exception
that applied to the entire file, but was only meant to target this block
of code.