🚧 filter: Add a --debug option
Now that disk space usage is uncapped (compared to previous pandas
implementation using in-memory chunks), it can be useful to know how
large the database file was.

However, I'll have to think more about how useful this is once database
files are passed in by the user. Ideas:

- Mark this as an experimental feature for `augur filter`, to be changed
  or removed with any version.
- Add it to the `augur db` interface, e.g. output of `augur db inspect`.
  However, it can still be useful to know the sizes of "intermediate"
  tables.

It'd also be useful to add runtime information here. Ideas:

- Print run times of each "major" function in real-time. This can
  probably be achieved by some sort of decorator function.