feat: allow reconstruction of amino acid sequences
augur ancestral so far only dealt with nucleotide sequences, while
translation were done codon by codon from the nucleotide sequences in
later steps. If translations for tips are available, the ancestral
amino acid sequences can be determined analogously via ancestral
reconstruction. This is implemented in this commit.

For each node, the node data structure will contain a `muts` and a
`aa_muts` field. The latter is a dict with gene/cds names as keys and
`<ancestral><pos><derived>` mutations (same as in augur translate).
In addition, an annotation and reference sequences are derived stored
in the node-data-json.