export v1: Remove --metadata-id-columns
This was added to both export v1 and v2 as part of a larger change¹, but
that change was incomplete and did not work as intended. It has been
fixed for export v2².

Since export v1 is no longer being maintained, remove the option instead
of trying to fix it here as well.

¹ 42f6cb2c8e23f496e56a2e803bb222b5076c1db3
² 9bf1b4ce4169b2a10ea974608188c2fdd918b9bc