Always expect "ref" to be a string
Fixes a bug with VCF inputs for augur ancestral where the reference
sequence is already a string, so it could not be converted to a string
from a BioPython sequence as the existing code expected. This commit
converts the BioPython Seq instance into a string after loading the root
sequence, so we always work with a string. It also updates the docstring
for `ancestral_sequence_inference` to include the "ref" argument.