Add functional test for root sequence and fix bug
Adds a functional test for the root sequence input of the new ancestral
interface and fixes a bug revealed by this test where the reference
sequence needs to be converted to a BioPython SeqIO.Seq instance prior
to calling the feature extraction method on it. The feature extraction
method accepts Seq and str inputs, but it will return an output of the
same type. When we previously passed a string, the returned string
didn't have the expected `translate` method.