filter: Remove attempt at extracting variables from --query
The attempt worked under limited testing, but introduced a bug that can
only be resolved with further complex Pandas query parsing. I don't
think that road is worth pursuing at the moment, so it's better to drop
the effort entirely.

This partially reverts commit 2ead5b3e3306dc1100b49eb774287496018122d9
and related changes.

I chose to keep the definition of PandasUndefinedVariableError at the
top-level¹ because it keeps external references next to each other, and
typing_extensions in the dependency list² because it is bound to be
useful at some point, and updating dependencies on Augur's Bioconda
recipe is a hassle.

¹ 602e3d5429f26d4625ecaeb7a07b47477f333d37
² 26586595e1b16a8ba77e0ed1bb7b83c766bddcc1