Update annotations schema to be correct
The annotations schema is updated to more accurately represent how this
data is used by Auspice version 2.46.1. Specifically, the strand of the
'nuc' property is not used by auspice, and start/end coordinates are
required properties.

Tests are updated because the tests were incorrect.

Currently the annotations section of node-data JSONs are produced by
`augur ancestral` and `augur translate`. Each adds a 'nuc' property,
and it's required for Auspice, so I think it's appropriate to make it a
required property for the schema.

The changes to the schema here are only reflected when parsing node-data
files, the next commit will make this available for our typical auspice
JSON validation.