Export clock rate covariance matrix and std dev
Exports the clock rate covariance matrix and std dev from the
phylogenetic regression performed by TreeTime in augur refine. This
information was already available in a TreeTime data structure and it is
useful to users who do not know the std dev of their pathogen's clock
rate already.

Adds functional tests for behavior when no clock rate is provided by the
user and when one is. In the latter case, we shouldn't have a covariance
matrix.