filter: Don't apply numerical conversion to columns not in query
Converting queried columns should be much better than converting
all columns: in most cases, only a small subset of metadata
columns are used in the query.

A stretch goal would be to convert only the columns used for numerical
comparison. However, detecting which are used *as numeric* is a harder
problem, and benefits are marginal for common queries that don't use
many columns.

Note that the column extraction function isn't perfect and it returns
all or nothing. If nothing is returned, then fall back to the previous
behavior of converting all columns.