feat(mask): allow empty mask file, warn not error
This commit adds a warning message when encountering an empty mask file. Now, instead of throwing an error, a warning is printed with the name of the empty mask file.

Reason: When stubbing out a build, one may want to add an empty mask file
to future proof the repo even if that list of sites is empty

Also: When one just has one masked site, removes it from the file, one generally doesn't want the workflow to error

Config file changes, e.g. removing mask sites, shouldn't cause workflows to error.
This is arguably a bug.