io: Add class interface for working with metadata files
Metadata is defined by its delimiter, columns, and rows. These
properties are inferred and returned in read_metadata, but that function
is pandas-focused. This commit is part of a PR with aims to move away
from using pandas to represent and access metadata at least in some
parts of the code. Alternatives will need a way to access these
properties, and this new class interface serves that purpose.

It will be used in subsequent commits.