filter: Stream filtered metadata to outputs directly
Previously, the output metadata and strain files were constructed as
multi-stage exports of an in-memory DataFrame representation of the
original metadata. This worked fine, except:

1. The order of rows was not guaranteed to follow that of the original
   metadata.
2. This relies on perfect round-tripping of file→DataFrame→file
   contents.
3. This relies on having all columns of the metadata available in the
   in-memory DataFrame representation.

(1) is not a big deal, but is addressed by this change.
(2) can be prone to issues, though nothing blocking has been brought to
our attention.
(3) is the main motivator given recent ideas to optimize I/O operations
in augur filter. With this change, there is no longer a need to have all
columns of the metadata stored in memory, which opens the door to
further optimizations.