Root sequence exported in main JSON
For smaller genomes it's a nicer experience to inline the root sequence
data to avoid having to manage an extra sidecar file, with the downside
being an (often negligible) increase in file size. Data structure
(schema) is identical between a root-sequence JSON and the top-level
JSON key `root_sequence`

No sanity checking is done on the genome size, but the intention is for
only small genomes to be inlined.

For full context, see discussions in Slack including
<https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1690930305357089>

Related Auspice PR <https://github.com/nextstrain/auspice/pull/1688>