[schema] Enforce some distance measure on trees
Trees could currently be produced with neither "div" nor "num_date"
information. Arguably Auspice could interpret these as cladograms
but as it stands these datasets aren't rendered by Auspice.

Datasets without this information are easy to create but rare in
practice: If the node-data files don't define "mutation_length" or
"branch_length" then there's no "div" and if they don't define
"num_date" then that's not there either.

Note that what we really want to require is that "div" is present on all
nodes and/or "num_date" is present on all nodes, but the schema doesn't
let us do this.