Remove tests under tests/builds
There are a few data files that are referenced by pytests. Move those to
tests/data.

These could be converted to Cram tests, though that would require some
work.

Also, note that the TB build removed here is not the same as the TB
build running in pathogen-repo-ci. I did a quick spot-check and noticed,
at the very least, the latter does not run `augur reconstruct-sequences`
while the former does. That's one less point of test coverage. Maybe it
is covered by other pathogen-repo-ci, but those aren't factored into the
code coverage report.