Export ancestral AA sequences for tree's root node
Fixes a bug in the new augur ancestral interface for amino acid
sequences where the output JSON was missing an entry for "aa_sequences"
for the root node that downstream tools like `augur clades` relied on.
The `augur translate` output creates this entry [1], so workflows that
use the nucleotide `ancestral` followed by the `translate` command would
not be affected. Workflows that use the new `ancestral` command would be
affected in cases where `augur clades` looks for amino acid mutations
that only exist in the root node and not on any subsequent
branches (e.g., for seasonal influenza B/Victoria's NA tree).

[1] https://github.com/nextstrain/augur/blob/b299059b38b1f579a70129c295405a19eb3f7c06/augur/translate.py#L291-L295