feat: add `clades2` command that works like traits
Currently, to assign clades at internal nodes, we require a `clades.tsv`
which isn't a natural way to define clades in most cases

Rather, one often wants to infer clades from a set of labeled tips

This PR adds a new subcommand `clades2` to infer clades at all tips and internal nodes
from a subset of labeled tips

Currently, the implementation is a stripped down version of `traits`
with confidence/entropy/model output removed
But this is an implementation detail that might change in the future

Some limitations of the current approach:
- clades can be non-monophyletic
- hierarchy information is not taken into account (i.e. the internal node at a junction of A, A.1.1 and A.1.2 will not be A.1 but one of the other three)

In theory, hierarchy could be taken into account but this may not be required for this command to be useful already.

The name is provisional, I couldn't immediately think of a good one so I went for `clades2`.