[ancestral] add general (simple-genome) tests
While we had tests that (probably?) covered the code here, using a
simple genome is much easier to reason with and to validate every
mutation. In my opinion it's also important to see tests using "big"
datasets (e.g. those bigger than we can manually verify each and every
mutation) as not necessarily testing the accuracy of the command, but
rather testing for regressions / changes in behaviour.

VCF-input tests using the same data will be added in a future commit
because (spoiler!) there are bugs which need fixing.