[ancestral] Add VCF tests
The added test here fails under TreeTime 0.11.1 but has been fixed as
part of <https://github.com/neherlab/treetime/pull/263>.

Description of the bug (as it relates to the test added in `vcf.t`):

The SNPs at nt 33 are encoded in the VCF as:
1	    33	.   A	C,G	.	    .	    .	    GT	    1	        2	        0
where ALT 1 ("C") is on Sample_A and ALT 2 ("G") is on Sample_B.
The ALT 2 is not being parsed by `read_vcf`, which results in
a changed mutation profile at pos 33:
.       **FASTA input**               **VCF input**
.         |---G33C-- sample_A           |---A33C-- sample_A
.  --A33G-|                      -------|
.         |--------- sample_B           |--------- sample_B
.
Because of this bug, the following test fails.

The `read_vcf` function is used in augur commands ancestral, refine,
sequence-traits, translate and tree.