[utils] lift error handling to load_features
Both functions which call this would check for a return value of None
(indicating that the file didn't exist) and cause the augur function to
exit. It's cleaner to lift this into `load_features` and this makes it
easier for that function to raise errors in the future (e.g. on
malformed/empty reference files).