[utils] always extract 'nuc' annotation (GFF)
This fixes a long-standing oversight in Augur where GFF files could only
define the nucleotide coordinates via a row with GFF type 'source'. We
now parse the (preferred) GFF type 'region' as well as the
'##sequence-region pragma'. This allows us to exit if the nuc
coordinates are not defined, and the error message should help users
correct their GFF files.

Note that the code is not yet implemented which guarantees the extracted
nuc coordinates will be exported in the JSON.

The changes here resulted in a number of tests needing updated, largely
due to us now parsing an additional feature from GFF files (the 'nuc'
feature)