[translate] guarantee nuc annotation produced
This builds off the preceding 3 commits which guarantee that a 'nuc'
feature will be parsed from the reference file. We now guarantee it'll
be exported in the node-data JSON.

Note that the change to the TB aa_muts.json test file was due to a bug
in the previous code, where `'type: feat['type']` would incorrectly
reuse the last defined `feat` in the preceding loop. (I think this is a
pitfall of using large "real-life" test files as it's impractical to
manually check the source-of-truth we are comparing against.)

Since the 'nuc' feature is guaranteed to exist, we can also check it
against the existing nuc annotation within `augur ancestral`, where
applicable.

Closes #953, although there is good commentary in that issue about
improving our parsing of GFFs more generally than that implemented here.

Closes #1346