Limit concurrent CI runs
Concurrent runs often happen due to iterative pushes to a PR. In those
cases, only the most recent commit is useful for CI to run.

Given current concurrency limitations, I've been in situations where the
most recent commit to a PR was waiting for CI runs on older commits to
finish before it could even start, due to the FIFO nature of GitHub
Actions workflow run scheduling. In those situations, I've manually
canceled the older CI runs to allow the most recent CI to run. This
change automates that process.

Concurrent CI runs on the default branch are useful, so those are still
allowed.