docs: Replace "breaking changes" with DEPRECATED
I created DEPRECATED without knowing that there was already a similar
page in the docs. That page was not very visible to developers, with no
mention in the dev docs.

Instead of continuing with the breaking changes file, move relevant
contents into DEPRECATED and automatically render DEPRECATED in the
docs.