Don't support Biopython 1.82
Biopython 1.82 was released a few hours ago¹ and has an undocumented
breaking change in the SeqFeature API which is used indirectly in augur
translate (through bcbio-gff).

It's possible that this should be a hard ceiling instead of a one-off
exclusion, but make it the latter for now. Fixing compatibility requires
upstream changes in bcbio-gff which is out of our control.

¹ <https://github.com/biopython/biopython/blob/biopython-182/NEWS.rst>