setup-micromamba: Invalidate cache daily
The default cache key is dependent on changes in usage of this action,
and does not invalidate based on time. Since we don't change usage of
this action often, this leads to long cache times, which are not
desirable.

Defining a time-based cache expiry can be subjective, but I'll start
with daily - this is the easy to implement, and only results in one
longer running CI workflow per day.