[ancestral] fix ambiguous site behaviour (FASTA)
Fixes a bug where positions which were all ambiguous (i.e. masked)
positions were being reported as mutations in the JSON output.

Furthermore, while the mask was being applied to sequence export to make
all masked positions "N", if we are inferring ambiguous sites then we
should use the reference base instead. This is now fixed.

Closes #1382 for the FASTA-input case only.

Unfortunately the added test doesn't actually recreate the first bug -
for this simple example TreeTime is not inferring a ATGC allele at the
all-Ns position. I can recreate this (and discovered it) using Cholera
data. I presume this is related to numerical consistency across the
(flat) likelihood values in this small test case.

Variable names are changed: `root_sequence` to `reference_sequence`.
This is much clearer (to my eyes at least).

NOTE: A further bug was discovered while working on this, which involved
behaviour when we are not inferring ambiguous bases. See the added test
here, which is currently disabled as otherwise CI fails. This bug is not
newly introduced by this work.