frequencies: Read all metadata as string type
Metadata is used for strain and date columns, which are already read as
string. Other columns are used for KDE estimation. Skipping dtype
inference on other columns should result in faster reading of metadata.

A more optimal change would be to skip loading those unused columns
entirely, but that requires another change in read_metadata.