Parameterize the record ID field for output sequences
Include a `--output-id-field` parameter in augur parse to indicate the ID
field for the output sequences file, such as using "accession" instead of
"strain".

It is important to note that the `--output-id-field` parameter is not required,
and augur parse will fall back to the DEFAULT_ID_COLUMNS (e.g. ('strain','name'))
if it is not present. If none of the DEFAULT_ID_COLUMNS are present in the fields,
fall back to using the first field.

The `--output-id-field` parameter is designed to accept a single field name,
not multiple. User are required to provide a `--fields col1 col2 strain accession`
argument in the same invocation. It seems reasonable to expect the user to
choose a specific field name for the ID.

To prevent unintended behaviors, if `--output-id-field` is not present in
`--fields` (e.g., due to a typo), augur parse will raise an error instead of
falling back to DEFAULT_ID_COLUMNS.