Refactor and reuse the metadata default ID columns instead of hardcoding
During augur parse, instead of hardcoding a check for a 'name' field and
then a 'strain' field to use as a sequence ID, reuse and check against the
io.metadata DEFAULT_ID_COLUMNS list to be more consistent with the rest of
augur.

This will result in a breaking behavior change in that parse is now checking
for 'strain' before 'name' as the default ID column. However, the argument
to be more consistent with the rest of augur is that the DEFAULT_ID_COLUMNS
list is already used in other parts of augur and augur parse should be no
different.