export v2: Add `metadata_columns` to Auspice config schema
Since the new `--metadata-columns` flag controls
what data is available in the final Auspice JSON,
it should have a matching option in the Auspice
config JSON setting.

This also moves the flag to the "DISPLAY CONFIGURATION"
group to match the other Auspice config controlled
settings. As per the existing pattern, the command
line flag overrules what is set in the config file.

Based on feedback from @jameshadfield
https://github.com/nextstrain/augur/pull/1384#discussion_r1468988767