Define a constant for shared pandas.read_csv() options
Start with just an explicit encoding. Applied to all invocations of
read_csv().

The current motivation of this change is to enforce UTF-8 encoding.
Another way to address that motivation would be to use Augur's internal
open_file() in place of the first argument to read_csv() to handle
encoding options as well as supported compression formats - but that's
not a trivial change.