Initialize variables in proper scope
Commit "Annotate tips with the minimum necessary information"
(dde857a23d) referenced the weights_attribute variable before it was
initialized, resulting in an UnboundLocalError when a non-KDE method
(i.e. diffusion) is used with --tree. This brings that variable and
another related one to a higher scope where the reference is now valid.