Use Pandas function to extract all names from query
Previously, ast.walk did not support the special case of backtick
quoting that is supported by pandas.DataFrame.query. Instead, use the
function that Pandas uses internally to parse queries.

This includes support for backtick quoting, but loses support to
distinguish variable names from other names as mentioned in the changes.
Intersecting the return value of this function with known metadata
column names should give the columns used in the query, which is the end
goal.