[ancestral, translate] validate output node-data
By leveraging the existing validation designed for reading node-data
files we not only validate against the annotation schema but also
perform other sanity checks on the node-data structure for which a
schema does not (yet) exist.

As part of this we add a `--validation-mode` argument following `augur
export v2` which allows users to skip validation of the output node-data
and (translate only) skip validation of the input node-data.