Validation mode warn no longer exits code 2
The help message for `--validation-mode` is very specific:

> 'error' mode causes a non-zero exit status if any validation checks
> failed, while 'warn' does not.

In the actual implementation however we would print an error and exit
with code 2 for any validation check _except_ the augur version check.
Now, for `ValidationMode.WARN` we catch these errors and turn them into
warnings such that augur can exit code 0 as specified.

Note that during validation, any non-ValidationError exception will be
be fatal even in `WARN`. An example of this would be errors loading the
schema itself. See the following review comment for more:
<https://github.com/nextstrain/augur/pull/1440#discussion_r1531113449>