docs: Tone down warning about Debian package
I think the note about Debian's package was a bit harsh as written.  We
rely on a ton of Debian's packages for our own work and they're
typically very high quality for a package ecosystem.  Removing "Use at
your own risk" as that implies some vague fear it may be bad that I
don't think is justified.  Instead, spell out that we don't recommend it
and why.