Deduplicate augur refine docs
Having a copy of the CLI reference on the FAQ page can be handy, but a
link to the actual reference page is sufficient. A similar argument can
be made for moving the guide out of the reference page.

Aligns with recent changes in "Link to new location for
filtering/subsampling docs" (8bdb0be2cd)