Improve speed of checking duplicates
pandas's isin(set) converts the set into a list, making each value check
scale linearly with the size of the set.¹ In the usage here, it's
unfortunate because the size of the set grows with each metadata chunk
processed. That means the entire operation scales with (size of chunk *
size of metadata).

Do the equivalent but much faster by first making a set from the current
chunk's index then intersecting that with the set of previously seen
IDs. This is faster because both of these steps scale linearly² with the
size of the chunk, not the size of the metadata, meaning the entire
operation scales with (2 * size of chunk).

¹ <https://github.com/pandas-dev/pandas/issues/25507>
² <https://wiki.python.org/moin/TimeComplexity#set>