Skip annotations for fields that do not exist in record
I don't think we should support adding new fields with record
annotations because it's hard to ensure that the new field gets added
across all records that are streamed through the command.

This is a deviation from the original nextstrain/ingest script, but I
think a necessary one to avoid bugs in the final output metadata.