[export] restrict sig figs for confidence/entropy
Using fewer sig figs (or decimal places, as appropriate) is just fine
for Auspice's usage and helps reduce the output JSON size. Testing on
the zika dataset this reduces the (minified, gzipped) JSON by 20% from
175kB to 139kB.

This refactor also uses more thorough error checking and enforcement
that entropy values are found together with confidence values where
appropriate. (Note that the previous usage of `is_valid` was bogus as it
only works for string values.)