Run mypy as a separate job on a single Python version
Previously, mypy was run under pytest. This was easier to set up, but
we've since realized that it's not necessary or practical to run mypy on
various Python versions in CI. Instead, it should be run on the minimum
supported Python version to ensure types are usable by all Python
versions that Augur is compatible with.