Fix location field recognition [#1530]
Previously, if the `geo-loc-name` field was the empty string, the `if
not location_field` condition would match, and the early return
branch would trigger.

When processing multiple records, some of which had empty
`geo-loc-name` fields and some of which didn't, this command would
produce output where some records had `country`, `division`, and
`location` fields added and some didn't, which triggers the
"inconsistent fields" check in the top-level `run()` method, causing
the command to exit with an error.

Instead, use a default value of `None` when trying to get the
`geo-loc-name` value from the record, and then check explicitly for
`None` in the "early out" condition.

Also adds a test to validate this works as expected.