curate: Accept Excel and OpenOffice metadata files
Reading Excel spreadsheets directly will be helpful for users (and
ourselves!) to avoid repeated manual conversions and ad-hoc Excel → TSV
programs.  OpenOffice supports comes along for the ride with Calamine,
the Rust library we're using (via unofficial Python bindings) to read
the files.

Only the first sheet in the workbook is read, and how it's read is not
configurable.  In review of previous Excel files we've used, this is
sufficient functionality.  We can extend support across alternate sheets
or multiple sheets or parts of sheets in the future, if needed.

Using Calamine means that we can support .xls and .xlsx with a single
library instead of using both xlrd and openpyxl.  The latter is
certainly doable—I've done it that way many times in the past; it's also
what Pandas does—and our needs a pretty basic so they'd suffice, but
still, a zero deps, single library, Rust-underneath solution is nice,
and I was pleased to find it.  If we find out Calamine is lacking in
some way, we can still fallback to the tried and true xlrd + openpyxl
combo.

Resolves: <https://github.com/nextstrain/augur/issues/1487>