CI: Update pathogen-repo-ci
Add comments to clearly show the job `pathogen-repo-ci-v0` is for
pathogen repos that do not conform to the standard pathogen repo
structure. Since these are tied to an old structure of pathogen repos,
the job should not be expected to be updated.

Any new pathogen repos should be added to the `pathogen-repo-ci`
job which replicates the centralized `pathogen-repo-ci` GH workflow.
Replicating the centralized `pathogen-repo-ci` because we've made the
decision to _not_ support ambient Augur version¹

¹ <https://github.com/nextstrain/.github/issues/66>