curate: Read the first *visible* *worksheet* from Excel/OpenOffice files
Sheets may be hidden (or even "very hidden"), and we'll skip those.
Sheets may also not be worksheets but be chart sheets, macro sheets, and
more, so we'll skip those as well.  Hidden sheets seem more likely to
crop up in practice, but we're filtering anyway so it's not much more
effort to be precise about the sheet type too.

Additionally, clarify in documentation (and code) the behaviour of
skipping initial empty rows/columns in the sheet, as @joverlee521 made
the good suggestion to do in review.