Fix titer substitution model tree annotations
Fixes a bug in the `augur titers sub` tree annotations in the JSON
output where antigenic weights were assigned per branch for the opposite
substitution values than expected. Specifically, the bug caused each
branch to get the antigenic weight associated with substitutions in the
form of "<derived allele><position><ancestral allele>" instead of
"<ancestral allele><position><derived allele>". For example, when the
substitution model found antigenic weights for HA1 substitutions N193S
and S193N, the tree annotations would have assigned the N193S weight to
branches with the S193N substitution.

This commit expands an existing functional test to check for a
data-specific antigenic weight. After confirming this test failed, I
fixed the bug and confirmed that the test passed.

Note that this bug dates back to December 2018 when I first added the
code to annotate the tree with the substitution model weights [1]. This bug
had no impact on the accuracy of the underlying model itself, so we
would never have noticed it without manually comparing the titer drops
in the tree to the "substitution" array in the JSON output.

[1] https://github.com/nextstrain/augur/commit/6721b7d67193361b9b41906e7cbc1d0346289350#diff-96cfc90794aa092cbb7b8577ca92d5757a777325d78078da3761ea26ee4c5956R67