merge: Support table-specific id columns and delimiters
When trying out `augur merge` for our common Nextclade metadata merge
step¹, I realized it would be good to allow table-specific values for
these settings.

Table-specific values avoid problems with impossible orderings, e.g. two
files both have columns A and B but you want to use A as the id column
in the first and B in the second.  Maybe that's an edge case, but it's
also just much easier to reason about table-specific values than reason
about how an ordered list of N values applies to all M files.  A shared
list of values is still good for when you don't know what they'll be
ahead of time.

¹ <https://github.com/nextstrain/pathogen-repo-guide/blob/89b3c5dbc9b8f6a009f4a19c3ac56113bc5511ee/ingest/rules/nextclade.smk#L75-L95>