docs: Improve AUGUR_DEBUG description
Includes clarification of "a truthy value" as "a non-empty value", since
all env vars values are strings and any non-empty string is a truthy
value in Python.  The new phrasing matches existing phrasing in our docs
and makes it much clearer to a reader not thinking about details of
Python's values.