io.metadata: Improve specificity of a "no id column found" error message
This message in Metadata() now matches the error message for the same
condition in read_metadata(), which is a big improvement to the "say
what now??" quotient.  It seems like Metadata's _find_first()
implementation was intentionally made generic thus requiring a generic
error message, but it's only used in a single place, to find the id
column, so make it specific and thus clearer.