Show message from pandas UndefinedVariableError
The goal is to show the name of the missing column in the error message.
Something like this would be ideal:

    ERROR: Query contains a column 'invalid' that does not exist in metadata

However, UndefinedVariableError does not expose the column name apart
from being in the error message text. Showing the entire message text is
slightly redundant, but seems better than parsing the column name out of
message text that's subject to change.