merge: Ignore the user's ~/.sqliterc file, if any
By default that file is evaluated by the SQLite CLI when it starts, and
the user may have commands in it that affect default CLI behaviour we
rely upon, such as .mode.  Since it's hard to know exactly how to reset
to a "pristine" settings state, avoid the user's rc file entirely.

Resolves: <https://github.com/nextstrain/augur/issues/1603>