Preserve formatting of command descriptions
This is useful for commands such as `augur distance` which have
descriptions written in rST. The rST syntax is preserved and could be
removed with a custom formatter class (as done in Nextstrain CLI¹), but
simply using the built-in RawDescriptionHelpFormatter is an improvement
over the default behavior.

Note that this only applies to descriptions, not help text for options.

¹ https://github.com/nextstrain/cli/blob/d4a419626eb5dc64e6de08566e51c5b48814727d/nextstrain/cli/argparse.py#L53