curate format-dates: Move `--expected-date-formats` to optional args
Because of the added default values in
<https://github.com/nextstrain/augur/pull/1501>, this is no longer a
required argument.

The alternative is to explicitly mark this option as `required=True`,
but that's a breaking change that can be considered later. Currently,
if `--expected-date-formats` is not provided and dates match the
defaults, then it's a no-op. If dates do not match the default formats,
it will raise a loud error so the user can add the custom formats.