feat(ancestral): Add optional Nextclade GFF3 compatibility mode
Until this PR, it was hard to use augur ancestral with a genome annotation gff3 from a Nextclade dataset for multiple reasons:
- The GFF reader in augur looks for "gene" features only, while Nextclade 3 uses primarily CDS (gene only for backwards compatibility)
- Augur's GFF reader doesn't support compound features
- It takes the "gene"/"feature" name in the order "gene" > "gene_name" > "locus_tag", which differs from Nextclade's order

This meant that users had to create a Genbank file, just for the purpose of running augur ancestral.
That was doable but tedious.

This PR solves the problem by adding a new GFF reader mode that is designed to be as compatible with Nextclade as possible. That means:
- It looks for CDS features (for simplicity it ignores Nextclade's gene backwards compatibility)
- It supports compound features
- It takes the feature name in the exact same order as Nextclade

The new mode is activated by passing the `--nextclade-compatibility` flag to augur ancestral.

It does no harm to anyone who does not want to use it.

It results in some code duplication, but I think it's totally worth it for the sake of usability.
I have already started using it in some of my private builds and it works like a charm.