release: Remove condition on event_name
This was flawed: even though ci.yaml is called by another workflow, the
`github` context is associated with the caller workflow¹ so event_name
is "workflow_dispatch" during a release run.

It doesn't make sense to update the condition to match
"workflow_dispatch", since there are valid reasons for running ci.yaml
using workflow_dispatch.

I think it makes more sense to simply remove this half of the condition
as redundant². The other half of the condition sufficiently ensures that
the job is run only when called by the release workflow on the default
branch.

¹ <https://docs.github.com/en/actions/sharing-automations/reusing-workflows>
² <https://github.com/nextstrain/augur/pull/1634#discussion_r1769243371>