Condition release job on default branch
This simplifies the dev docs with one less reference to a hardcoded
default branch name.

Note that it's still possible for this to run from another branch if the
condition is removed in that branch¹, but one would have to very
intentionally make that change *and* select the branch when dispatching.

¹ <https://stackoverflow.com/questions/74633548#comment133256716_746335499>